+++ ServeAll.Core/Entities/ReturnWareHouse.cs
@@ -11,7 +11,7 @@ namespace ServeAll.Core.Entities
         public string return_code    { get; set; }
         public int product_id        { get; set; }
         public string return_number      { get; set; }
-        public decimal return_quantity { get; set; }
+        public int return_quantity { get; set; }
         public int branch_id         { get; set; }
         public string destination      { get; set; }
         public DateTime return_date { get; set; }
